pyproject.toml
[project]
name = "ensembl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-caching>=2.3.1",
    "pandas>=3.0.0",
    "plotly>=6.5.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
